fix: webhooks. authentication.

+++ app/(dashboard)/notifications/actions.ts
@@ -1,7 +1,10 @@
 'use server';
 
 import prisma from '@lib/prisma';
-import { CountdownSetting, NotificationsFilterSettings } from '@app/(dashboard)/notifications/client';
+import {
+	CountdownSetting,
+	NotificationsFilterSettings,
+} from '@app/(dashboard)/notifications/client';
 import { Routes } from 'discord-api-types/v10';
 import env from '@env';
 import { REST } from '@discordjs/rest';
@@ -9,17 +12,22 @@ import { createWebhook } from '@lib/discord-api';
 import { revalidatePath } from 'next/cache';
 import { Logger } from 'next-axiom';
 
+import { isAuthorized } from '@lib/server-utils';
+
 const rest = new REST({ version: '9' }).setToken(env.DISCORD_BOT_TOKEN);
 const log = new Logger();
 
 export const updateFilters = async (
 	guildId: string,
-	settings: NotificationsFilterSettings,
+	settings: NotificationsFilterSettings
 ): Promise<void> => {
-	// remove countdown key
+	const authorized = await isAuthorized(guildId);
+	if (!authorized) {
+		throw new Error('Unauthorized');
+	}
 
 	const data = Object.fromEntries(
-		Object.entries(settings).map(([key, value]) => [key, Number(value)]),
+		Object.entries(settings).map(([key, value]) => [key, Number(value)])
 	);
 	await prisma.enabled_guilds.update({
 		where: {
@@ -33,8 +41,13 @@ export const updateFilters = async (
 
 export const addCountdown = async (
 	guildId: string,
-	settings: CountdownSetting,
+	settings: CountdownSetting
 ): Promise<void> => {
+	const authorized = await isAuthorized(guildId);
+	if (!authorized) {
+		throw new Error('Unauthorized');
+	}
+
 	if (settings.days < 0 || settings.days > 31) {
 		log.error('Invalid days', {
 			guildId,
@@ -90,8 +103,13 @@ export const addCountdown = async (
 
 export const removeCountdown = async (
 	guildId: string,
-	minutes: number,
+	minutes: number
 ): Promise<void> => {
+	const authorized = await isAuthorized(guildId);
+	if (!authorized) {
+		throw new Error('Unauthorized');
+	}
+
 	await prisma.notification_countdown.delete({
 		where: {
 			guild_id_minutes: {
@@ -106,48 +124,64 @@ export const removeCountdown = async (
 
 export const updateChannel = async (
 	guildId: string,
-	channelId: string,
+	channelId: string
 ): Promise<void> => {
-	await prisma.enabled_guilds.update({
-		where: {
-			guild_id: BigInt(guildId),
-		},
-		data: {
-			notification_channel_id: BigInt(channelId),
-		},
-	});
+	const authorized = await isAuthorized(guildId);
+	if (!authorized) {
+		throw new Error('Unauthorized');
+	}
 
-	const [webhookUrl, resp] = await Promise.all([
-		createWebhook(channelId, 'News'),
+	const [newWebhookURL, guild] = await Promise.all([
+		createWebhook(channelId, 'Notifications'),
 		prisma.enabled_guilds.findFirst({
 			where: {
 				guild_id: BigInt(guildId),
 			},
-			select: {
-				notification_webhook_url: true,
-			},
 		}),
 	]);
-
-	if (resp?.notification_webhook_url) {
-		await rest.delete(
-			Routes.webhook(resp.notification_webhook_url.split('/')[5]),
-		);
-	}
 
 	await prisma.enabled_guilds.update({
 		where: {
 			guild_id: BigInt(guildId),
 		},
 		data: {
-			notification_webhook_url: webhookUrl,
+			notification_webhook_url: newWebhookURL,
+			notification_channel_id: BigInt(channelId),
 		},
 	});
 
-	revalidatePath('/notifications');
+	if (guild?.notification_webhook_url) {
+		await rest
+			.delete(
+				Routes.webhook(guild.notification_webhook_url.split('/')[5])
+			)
+			// cleanup on error
+			.catch(async e => {
+				await prisma.enabled_guilds.update({
+					where: {
+						guild_id: BigInt(guildId),
+					},
+					data: {
+						notification_channel_id: null,
+						notification_webhook_url: null,
+					},
+				});
+				await rest.delete(Routes.webhook(newWebhookURL.split('/')[5]));
+
+				revalidatePath('/notifications');
+				throw e;
+			});
+	} else {
+		revalidatePath('/notifications');
+	}
 };
 
 export const disableFeature = async (guildId: string): Promise<void> => {
+	const authorized = await isAuthorized(guildId);
+	if (!authorized) {
+		throw new Error('Unauthorized');
+	}
+
 	const resp = await prisma.enabled_guilds.findFirst({
 		where: {
 			guild_id: BigInt(guildId),
@@ -160,9 +194,11 @@ export const disableFeature = async (guildId: string): Promise<void> => {
 	console.log(resp?.notification_webhook_url?.split('/')[5]);
 
 	if (resp?.notification_webhook_url) {
-		await rest.delete(
-			Routes.webhook(resp.notification_webhook_url.split('/')[5]),
-		);
+		await rest
+			.delete(Routes.webhook(resp.notification_webhook_url.split('/')[5]))
+			.catch(e => {
+				console.error(e);
+			});
 	}
 
 	await prisma.enabled_guilds.update({
